combines release logic into relay service
the relay and release services are both thin services that offlooad most logic
to either the node or api

combining the logic of the release service into the relay service simplifies the
infrastructure required for the bridge
